+++ clients/all.d.ts
@@ -128,3 +128,4 @@ export import Cloud9 = require('./cloud9');
 export import ServerlessApplicationRepository = require('./serverlessapplicationrepository');
 export import ServiceDiscovery = require('./servicediscovery');
 export import WorkMail = require('./workmail');
+export import AutoScalingPlans = require('./autoscalingplans');