+++ clients/all.js
@@ -129,5 +129,6 @@ module.exports = {
   Cloud9: require('./cloud9'),
   ServerlessApplicationRepository: require('./serverlessapplicationrepository'),
   ServiceDiscovery: require('./servicediscovery'),
-  WorkMail: require('./workmail')
+  WorkMail: require('./workmail'),
+  AutoScalingPlans: require('./autoscalingplans')
 };