+++ lib/config_service_placeholders.d.ts
@@ -130,6 +130,7 @@ export abstract class ConfigurationServicePlaceholders {
   serverlessapplicationrepository?: AWS.ServerlessApplicationRepository.Types.ClientConfiguration;
   servicediscovery?: AWS.ServiceDiscovery.Types.ClientConfiguration;
   workmail?: AWS.WorkMail.Types.ClientConfiguration;
+  autoscalingplans?: AWS.AutoScalingPlans.Types.ClientConfiguration;
 }
 export interface ConfigurationServiceApiVersions {
   acm?: AWS.ACM.Types.apiVersion;
@@ -262,4 +263,5 @@ export interface ConfigurationServiceApiVersions {
   serverlessapplicationrepository?: AWS.ServerlessApplicationRepository.Types.apiVersion;
   servicediscovery?: AWS.ServiceDiscovery.Types.apiVersion;
   workmail?: AWS.WorkMail.Types.apiVersion;
+  autoscalingplans?: AWS.AutoScalingPlans.Types.apiVersion;
 }